Avoid circular dependency issue on 1.10 with Random
